fixed a bad rename from fish back to fishes for consistency. Fixed database hookups
